GO-6630 Rearrange files
